feat: dto 타입 변경

+++ app/src/main/java/woowacourse/shopping/remote/Pageable.kt
@@ -4,7 +4,7 @@ data class Pageable(
     val sort: Sort,
     val pageNumber: Int,
     val pageSize: Int,
-    val offset: Int,
+    val offset: Long,
     val paged: Boolean,
     val unpaged: Boolean,
 )
